requirements.txt
pluggy>=0.13.1
Pillow>=10.2.0
pygame>=1.9.6
pygame-menu==4.0.7
pygame-vkeyboard>=2.0.8
psutil>=5.5.1
gpiozero>=1.5.1
picamera2>=0.3
opencv-python
numpy
pytest
